health-first-server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
django-ratelimit==4.1.0
bcrypt==4.1.2
celery==5.3.4
redis==5.0.1
flower==2.0.1
Pillow==10.1.0
django-extensions==3.2.3
python-dotenv==1.0.0
phonenumbers==8.13.25
validators==0.22.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
drf-yasg==1.21.7
django-filter==23.5 